- Fixed tensorflow datatypes in feature_description

+++ wrecksys/data/datasets.py
@@ -112,10 +112,10 @@ class TensorflowRecords(WrecksysDataset):
         files = [str(f) for f in self.output_dir.glob('*.tfrecord')]
         feature_description = {
             'context_id': tf.io.FixedLenFeature(
-                [self.max_length], tf.int32, default_value=np.repeat(0, self.max_length)),
+                [self.max_length], tf.int64, default_value=np.repeat(0, self.max_length)),
             'context_rating': tf.io.FixedLenFeature(
                 [self.max_length], tf.float32, default_value=np.repeat(3, self.max_length)),
-            'label_id': tf.io.FixedLenFeature([1], tf.int32)
+            'label_id': tf.io.FixedLenFeature([1], tf.int64)
         }
 
         def _parse(example):
